move makeCallback to separate module
