+++ src/routes/user.ts
@@ -1,30 +1,12 @@
 import { Router } from 'express';
 
 import { getUserController } from '../controllers';
-import { RequestEntity } from '../models/http';
 import { authenticate } from '../middleware/auth';
+import { makeCallback } from '../utils/express-utils';
 
 const router = Router();
 
 let userController = getUserController();
-
-const makeCallback = (handler: (body: RequestEntity<any>) => Promise<any>) => {
-    return async (req, res, next) => {
-        let body = {
-            url: req.url,
-            body: req.body,
-            params: req.params,
-            user: req.user,
-        };
-        try {
-            let response = await handler(body);
-            return res.status(response.statusCode)
-                .json(response.body);
-        } catch (e) {
-            next(e);
-        }
-    }
-}
 
 router.post('/', makeCallback((req) => userController.postUser(req)));
 router.post('/authenticate', makeCallback((req) => userController.authenticateUser(req)));
